services/api/pyproject.toml
[project]
name = "dinamo-marketing-api"
version = "1.0.0"
description = "GNK Dinamo Zagreb - AI-Powered Marketing Platform"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
    # Task queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    # HTTP client
    "httpx>=0.28.0",
    # AI
    "anthropic>=0.40.0",
    # Config
    "pydantic-settings>=2.6.0",
    # Utilities
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "orjson>=3.10.0",
    # PDF generation
    "reportlab>=4.2.0",
    # Date handling
    "python-dateutil>=2.9.0",
    # Trends
    "pytrends>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
